Plot preview and actual in units of nanosecond-Volt
Before sending the waveform it is useful to know the timescale and amplitude of
the pulses, in other words, what the waveform should look like before being
sent out. Then it can be compared with the actual oscilloscope data which is
plotted on in those same units. In this case, its easy to see the input
waveform and measured waveform have pulses about 5 nanoseconds FWHM.